+++ workspace/lib/mix/tasks/workspace.test.coverage.ex
@@ -217,7 +217,7 @@ defmodule Mix.Tasks.Workspace.Test.Coverage do
 
     paths =
       workspace
-      |> Workspace.projects()
+      |> Workspace.projects(order: :alphabetical)
       |> Enum.filter(fn project -> !project.skip end)
       |> Enum.map(&cover_compile_paths/1)
 
@@ -245,7 +245,7 @@ defmodule Mix.Tasks.Workspace.Test.Coverage do
 
     project_statuses =
       workspace
-      |> Workspace.projects()
+      |> Workspace.projects(order: :alphabetical)
       |> Enum.filter(fn project -> !project.skip end)
       |> Enum.reduce([], fn project, acc ->
         {coverage, module_stats} =